% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{docx_template}
\alias{docx_template}
\title{Return path to U54 Word template}
\usage{
docx_template()
}
\value{
A path to the Word document template.
}
\description{
Return the path to a word template for reporting.
}
\details{
There is a common word document template used for all of the U54 reporting
via officedown. This template has the U54 banner and various settings preconfigured.
Rather than duplicate the word document many times, it is stored in a default directory
within the package installation.

This function returns the path to this file, so that knitr/rmarkdown can refer to the
template indirectly.
}
\examples{
docx_template()

}
\seealso{
\code{\link[=knit_docx_with_datestamp]{knit_docx_with_datestamp()}}
}
