% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{update_packages}
\alias{update_packages}
\title{Update package from github}
\usage{
update_packages()
}
\value{
Result of [\code{\link[devtools:remote-reexports]{devtools::install_github()}}]
}
\description{
Update package from github
}
\examples{
\dontrun{
update_packages()
}
}
